refactor: remove all-catch handler
This also catches exceptions raised with `flask.abort`, and in those
cases it does not make sense to prepend the exception type name.

Rather, error handlers as recently implemented in the cookiecutter[1]
should be used. In the meantime, removing this means uncaught exceptions
will now be tracked in Sentry/logs but not forwarded to the client.

[1] https://github.com/DD-DeCaF/cookiecutter-flask-microservice/pull/82